// frontend/src/app/core/api/model/class-student-response.ts
/**
 * Training school system Backend API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ClassStudentResponse { 
    firstName?: string;
    lastName?: string;
    phone?: string;
    studentCode?: string;
    enrollmentStatus?: ClassStudentResponse.EnrollmentStatusEnum;
    gender?: ClassStudentResponse.GenderEnum;
    enrollmentId?: number;
}
export namespace ClassStudentResponse {
    export const EnrollmentStatusEnum = {
        PreEnrolled: 'PRE_ENROLLED',
        Incomplete: 'INCOMPLETE',
        Waitlisted: 'WAITLISTED',
        ConditionallyValidated: 'CONDITIONALLY_VALIDATED',
        Validated: 'VALIDATED',
        Suspended: 'SUSPENDED',
        DroppedOut: 'DROPPED_OUT',
        Rejected: 'REJECTED',
        Cancelled: 'CANCELLED',
        Completed: 'COMPLETED'
    } as const;
    export type EnrollmentStatusEnum = typeof EnrollmentStatusEnum[keyof typeof EnrollmentStatusEnum];
    export const GenderEnum = {
        Male: 'MALE',
        Female: 'FEMALE'
    } as const;
    export type GenderEnum = typeof GenderEnum[keyof typeof GenderEnum];
}