// frontend/src/app/core/api/model/unassigned-enrollment-response.ts
/**
 * Training school system Backend API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UnassignedEnrollmentResponse { 
    enrollmentId?: number;
    enrollmentNumber?: string;
    firstName?: string;
    lastName?: string;
    studentCode?: string;
    birthDate?: string;
}